+++ kierki-serwer.cpp
@@ -172,7 +172,7 @@ int main(int ac, char *av[])
             // deal_wrapper.write(client_fd_sp->to_int(), game_master_sp->get_game_type(), game_master_sp->get_whose_turn(), game_master_sp->get_player_cards(new_p_position));
 
             ingame_comm_wrappers::TOTAL_Wrapper total_wrapper;
-            total_wrapper.write(client_fd_sp->to_int(), std::map<PlayerPosition, uint16_t> {
+            total_wrapper.write(client_fd_sp->to_int(), std::map<PlayerPosition, uint32_t> {
                 {N, 33333}, {E, 255}, {S, 1}, {W, 2137}});
 
             // std::thread t(
